Moving validators to separate file (#177)
